pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "skls-core"
version = "0.1.0"
authors = [
    {name = "SKLS", email = "author@example.com"},
]
description = "A comprehensive knowledge graph and language processing system"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    # Core dependencies
    "pydantic>=1.10.0",
    
    # Embedding dependencies
    "requests>=2.25.0",
    "chromadb>=0.4.0",
    
    # Generator dependencies
    "json_repair>=0.28.0",
    "pyparsing>=3.0.0",
    "tqdm>=4.60.0",
    "python-dotenv>=0.19.0",
    "google-generativeai>=0.3.0",
    
    # Neo4j dependencies
    "neo4j>=5.0.0",
    
    # Additional dependencies
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.8",
]
all = [
    "skls-core[dev]",
]

[project.urls]
Homepage = "https://github.com/yourusername/skls_core"
Repository = "https://github.com/yourusername/skls_core"
Issues = "https://github.com/yourusername/skls_core/issues"

[tool.setuptools.packages.find]
include = ["skls_*", "skls_*.*"]
exclude = ["tests*", "examples*"]

[tool.setuptools.package-data]
skls_core = ["*.md", "*.txt"]

[tool.black]
line-length = 88
target-version = ['py38']